real_estate_image_generation/CFWD/.ipynb_checkpoints/requirements-checkpoint.txt
absl-py==2.0.0
addict==2.4.0
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.2
docopt==0.6.2
einops==0.7.0
filelock==3.13.1
fsspec==2023.10.0
ftfy==6.1.1
future==0.18.3
google-auth==2.23.4
google-auth-oauthlib==1.1.0
grpcio==1.59.2
huggingface-hub==0.19.1
idna==3.4
imageio==2.32.0
importlib-metadata==6.8.0
lazy_loader==0.3
lmdb==1.4.1
Markdown==3.5.1
MarkupSafe==2.1.3
networkx==3.2.1
numpy==1.21.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
opencv-python==4.6.0.66
packaging==23.2
pandas==2.0.3
Pillow==9.2.0
pipreqs==0.4.13
platformdirs==4.0.0
protobuf==4.23.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
python-dateutil==2.8.2
pytorch-msssim==1.0.0
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0.1
regex==2022.8.17
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.4.0
scikit-image==0.21.0
scipy==1.10.1
six==1.16.0
tb-nightly==2.16.0a20231113
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tifffile==2023.9.26
timm==0.9.10
tomli==2.0.1
torch==1.13.1
torchvision==0.14.1
tqdm==4.64.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
wcwidth==0.2.10
Werkzeug==3.0.1
yapf==0.40.2
yarg==0.1.9
zipp==3.17.0
